strip out redundant modules
