--- ui/src/auth/landing-error.js
@@ -1,49 +1,0 @@
-/**
-@license
-Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-import '@polymer/polymer/polymer-legacy.js';
-
-import '@polymer/iron-flex-layout/iron-flex-layout.js';
-import './landing-button.js';
-import { Polymer } from '@polymer/polymer/lib/legacy/polymer-fn.js';
-import { html } from '@polymer/polymer/lib/utils/html-tag.js';
-
-Polymer({
-  _template: html`
-    <style>
-      :host {
-        display: block;
-        text-align: center;
-        color: var(--app-secondary-color);
-      }
-
-      iron-icon {
-        display: inline-block;
-        width: 60px;
-        height: 60px;
-      }
-
-      h1 {
-        margin: 50px 0 50px 0;
-        font-weight: 300;
-      }
-    </style>
-
-    <h1>[[route.__queryParams.msg]]</h1>
-    <landing-button>
-      <a href="/">Go to the home page</a>
-    </landing-button>
-  `,
-
-  is: 'landing-error',
-
-  properties: {
-    msg: String,
-  },
-});
