--- ui/src/auth/landing-network-warning.js
@@ -1,68 +1,0 @@
-/**
-@license
-Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-import '@polymer/polymer/polymer-legacy.js';
-
-import './landing-button.js';
-import { Polymer } from '@polymer/polymer/lib/legacy/polymer-fn.js';
-import { html } from '@polymer/polymer/lib/utils/html-tag.js';
-
-Polymer({
-  _template: html`
-    <style include="landing-button">
-      :host {
-        display: block;
-        padding: 40px 20px;
-        text-align: center;
-        color: var(--app-secondary-color);
-      }
-
-      iron-icon {
-        display: inline-block;
-        width: 30px;
-        height: 30px;
-      }
-
-      h1 {
-        margin: 50px 0 10px 0;
-        font-weight: 300;
-      }
-
-      p {
-        margin: 0;
-      }
-
-      landing-button {
-        margin-top: 50px;
-      }
-    </style>
-
-    <div hidden$="[[offline]]">
-      <h1>Couldn't reach the server</h1>
-    </div>
-    <div hidden$="[[!offline]]">
-      <iron-icon icon="perm-scan-wifi"></iron-icon>
-      <h1>No internet connection</h1>
-      <p>Check if your device is connected to a mobile network or WiFi.</p>
-    </div>
-    <landing-button>
-      <button on-click="_tryReconnect">Try Again</button>
-    </landing-button>
-  `,
-
-  is: 'landing-network-warning',
-
-  properties: {
-    offline: Boolean,
-  },
-
-  _tryReconnect() {
-    this.fire('try-reconnect', null, { bubbles: false });
-  },
-});
