--- ui/src/auth/landing-tabs-overlay.js
@@ -1,178 +1,0 @@
-/**
-@license
-Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-import '@polymer/polymer/polymer-legacy.js';
-
-import { Polymer } from '@polymer/polymer/lib/legacy/polymer-fn.js';
-import { flush } from '@polymer/polymer/lib/legacy/polymer.dom.js';
-
-const documentContainer = document.createElement('template');
-
-documentContainer.innerHTML = `<dom-module id="landing-tabs-overlay">
-  <template strip-whitespace="">
-    <style>
-      :host {
-        position: absolute;
-        display: none;
-        pointer-events: none;
-        opacity: 0;
-        top: 0;
-        right: 0;
-        bottom: 0;
-        left: 0;
-        transition-property: top, right, bottom, left, opacity;
-      }
-    </style>
-  </template>
-  
-</dom-module>`;
-
-document.head.appendChild(documentContainer.content);
-Polymer({
-  is: 'landing-tabs-overlay',
-
-  properties: {
-    /**
-     * The element the overlay should cover.
-     */
-    target: {
-      type: Object,
-      observer: '_targetChanged',
-    },
-  },
-
-  listeners: {
-    transitionend: '_onTransitionend',
-  },
-
-  created() {
-    this._lastTarget = undefined;
-    this._transitionsInFlight = [];
-  },
-
-  _targetChanged(newTarget, oldTarget) {
-    if (!this._transitionsInFlight.length) {
-      if (this._lastTarget) {
-        this._lastTarget.classList.remove('landing-tabs-overlay-static-above');
-      }
-      this.style.display = 'block';
-      this._move(oldTarget, newTarget);
-      this._lastTarget = this.target;
-    }
-  },
-
-  _onTransitionend(event) {
-    const index = this._transitionsInFlight.indexOf(event.propertyName);
-    if (index >= 0) {
-      this._transitionsInFlight.splice(index, 1);
-    }
-
-    if (!this._transitionsInFlight.length) {
-      this._moveComplete();
-    }
-  },
-
-  _moveComplete() {
-    if (this._lastTarget !== this.target) {
-      this._move(this._lastTarget, this.target);
-      this._lastTarget = this.target;
-    } else {
-      if (this._lastTarget) {
-        this._lastTarget.classList.add('landing-tabs-overlay-static-above');
-      }
-      this.style.display = 'none';
-    }
-  },
-
-  _move(oldTarget, newTarget) {
-    const from = oldTarget || newTarget;
-    const to = newTarget || oldTarget;
-    if (!from && !to) return;
-
-    const fromOpacity = oldTarget ? 1 : 0;
-    const toOpacity = newTarget ? 1 : 0;
-
-    flush();
-    const thisRect = this.getBoundingClientRect();
-    const thisStyle = window.getComputedStyle(this);
-    const fromRect = from.getBoundingClientRect();
-    const toRect = to.getBoundingClientRect();
-
-    if (
-      toRect.top === 0 &&
-      toRect.right === 0 &&
-      toRect.bottom === 0 &&
-      toRect.left === 0 &&
-      toRect.width === 0 &&
-      toRect.height === 0
-    ) {
-      this.style.transitionProperty = 'none';
-      this.style.opacity = toOpacity;
-      this._transitionsInFlight = [];
-      this.async(this._moveComplete);
-      return;
-    }
-    this.style.transitionProperty = '';
-
-    let top = parseFloat(thisStyle.top || '0') + (fromRect.top - thisRect.top);
-    let right = parseFloat(thisStyle.right || '0') - (fromRect.right - thisRect.right);
-    let bottom = parseFloat(thisStyle.bottom || '0') - (fromRect.bottom - thisRect.bottom);
-    let left = parseFloat(thisStyle.left || '0') + (fromRect.left - thisRect.left);
-
-    this.style.transitionDuration = '0s';
-    this.style.transitionDelay = '0s';
-    const startValues = [
-      (this.style.top = `${top}px`),
-      (this.style.right = `${right}px`),
-      (this.style.bottom = `${bottom}px`),
-      (this.style.left = `${left}px`),
-      (this.style.opacity = String(fromOpacity)),
-    ];
-
-    top += toRect.top - fromRect.top;
-    right -= toRect.right - fromRect.right;
-    bottom -= toRect.bottom - fromRect.bottom;
-    left += toRect.left - fromRect.left;
-
-    const durations = [0.2, 0.2, 0.2, 0.2, 0.2];
-    const delays = [0, 0, 0, 0, 0];
-    // Delay left / right transitions if element is left / right of the target.
-    if (fromRect.left < toRect.left && fromRect.right < toRect.right) {
-      delays[3] = 0.1;
-    } else if (fromRect.left > toRect.left && fromRect.right > toRect.right) {
-      delays[1] = 0.1;
-    }
-
-    // Delay top / bottom transitions if element is above / below the target.
-    if (fromRect.top < toRect.top && fromRect.bottom < toRect.bottom) {
-      delays[0] = 0.1;
-    } else if (fromRect.top > toRect.top && fromRect.bottom > toRect.bottom) {
-      delays[2] = 0.1;
-    }
-
-    const endValues = [`${top}px`, `${right}px`, `${bottom}px`, `${left}px`, String(toOpacity)];
-
-    const names = ['top', 'right', 'bottom', 'left', 'opacity'];
-    for (let i = 0; i < startValues.length; i++) {
-      if (startValues[i] !== endValues[i] && !(durations[i] === 0 && delays[i] === 0)) {
-        this._transitionsInFlight.push(names[i]);
-      }
-    }
-
-    this.async(() => {
-      this.style.transitionDuration = durations.map(x => `${x}s`).join(', ');
-      this.style.transitionDelay = delays.map(x => `${x}s`).join(', ');
-      this.style.top = `${top}px`;
-      this.style.right = `${right}px`;
-      this.style.bottom = `${bottom}px`;
-      this.style.left = `${left}px`;
-      this.style.opacity = String(toOpacity);
-    }, 1);
-  },
-});
