--- ui/src/auth/lazy-resources.js
@@ -1,28 +1,0 @@
-import '@polymer/paper-styles/paper-styles.js';
-import '@polymer/paper-material/paper-material.js';
-import '@polymer/iron-meta/iron-meta.js';
-import '@polymer/iron-iconset-svg/iron-iconset-svg.js';
-import '@polymer/iron-icons/iron-icons.js';
-import '@polymer/iron-icons/hardware-icons.js';
-import '@polymer/iron-icons/social-icons.js';
-import '@polymer/iron-icons/places-icons.js';
-import '@polymer/iron-icons/editor-icons.js';
-import '@polymer/iron-icon/iron-icon.js';
-import '@polymer/paper-icon-button/paper-icon-button.js';
-
-import './landing-icons.js';
-import './landing-snackbar.js';
-import './landing-zoom-in.js';
-import './landing-get-started.js';
-import './landing-footer.js';
-import './landing-person.js';
-
-import './landing-tabs.js';
-import './landing-tab.js';
-import './landing-set-password.js';
-import './landing-reset-password.js';
-import './landing-forgot-password.js';
-
-import './landing-network-warning.js';
-import './landing-not-found.js';
-import './landing-error.js';
